% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duelSelectBoxInput.R
\name{duelSelectBoxInput}
\alias{duelSelectBoxInput}
\title{duel select input with drag and drop}
\usage{
duelSelectBoxInput(inputId, leftLabel, rightLabel, leftChoices, rightChoices)
}
\arguments{
\item{inputId}{id of chooser input}

\item{leftLabel}{label for left hand side}

\item{rightLabel}{label for right hand side}

\item{leftChoices}{options for left hand side}

\item{rightChoices}{options for right hand side}
}
\value{
shiny widget input
}
\description{
duel select input with drag and drop
}
